# Improve string column dict encoding performance