### What changes were proposed in this pull request?
Improve writing performance of encoded string column. Speed up by 2x 

### Why are the changes needed?
<!--
Please clarify why the changes are needed. For instance,
  1. If you propose a new API, clarify the use case for a new API.
  2. If you fix a bug, you can clarify why it is a bug.
-->


### How was this patch tested? 

I tested it with clickhouse branch: https://github.com/ClickHouse/ClickHouse/pull/68591 

Query:
``` sql 
set output_format_orc_dictionary_key_size_threshold = 1;
select concat('gluten ', cast(rand()%1000 as String)) from numbers(10000000) into outfile 'dict.orc' truncate;
```

Before current optimization: 
```
10000000 rows in set. Elapsed: 2.794 sec. Processed 10.00 million rows, 80.00 MB (3.58 million rows/s., 28.63 MB/s.)
Peak memory usage: 9.07 MiB.
```

After current optimization: 
```
10000000 rows in set. Elapsed: 1.423 sec. Processed 10.00 million rows, 80.00 MB (7.02 million rows/s., 56.20 MB/s.)
Peak memory usage: 9.07 MiB.
```



<!--
If tests were added, say they were added here. Please make sure to add some test cases that check the changes thoroughly including negative and positive cases if possible.
If it was tested in a way different from regular unit tests, please clarify how you tested step by step, ideally copy and paste-able, so that other reviewers can test and check, and descendants can verify in the future.
If tests were not added, please describe why they were not added and/or why it was difficult to add.
-->