@taiyang-li 2024-08-22
I have created a pr with current changes to upstream: https://github.com/apache/orc/pull/2010